Migrate project to CRA
