+++ .eslintcache
@@ -1,0 +1,1 @@
+[{"D:\\project\\spell-card-convert\\spell-card\\src\\App.tsx":"1","D:\\project\\spell-card-convert\\spell-card\\src\\phaser\\scenes\\preloader.js":"2","D:\\project\\spell-card-convert\\spell-card\\src\\phaser\\components\\cardCollection\\cardDisplayZone.js":"3","D:\\project\\spell-card-convert\\spell-card\\src\\phaser\\components\\navigation.js":"4","D:\\project\\spell-card-convert\\spell-card\\src\\phaser\\scenes\\GameScenes\\cardSwitch.js":"5","D:\\project\\spell-card-convert\\spell-card\\src\\phaser\\components\\gameSetting\\settingMenu.js":"6","D:\\project\\spell-card-convert\\spell-card\\src\\index.tsx":"7","D:\\project\\spell-card-convert\\spell-card\\src\\phaser\\scenes\\GameScenes\\mainGame.js":"8","D:\\project\\spell-card-convert\\spell-card\\src\\phaser\\scenes\\cardCollectionScenes\\newDeck.js":"9","D:\\project\\spell-card-convert\\spell-card\\src\\phaser\\components\\naviButton.js":"10","D:\\project\\spell-card-convert\\spell-card\\src\\components\\SpellCard.jsx":"11","D:\\project\\spell-card-convert\\spell-card\\src\\phaser\\scenes\\GameScenes\\deckSelection.js":"12"},{"size":155,"mtime":1608519169096,"results":"13","hashOfConfig":"14"},{"size":925,"mtime":1608519462919,"results":"15","hashOfConfig":"14"},{"size":1949,"mtime":1608519335709,"results":"16","hashOfConfig":"14"},{"size":749,"mtime":1608519700892,"results":"17","hashOfConfig":"14"},{"size":1114,"mtime":1608515685829,"results":"18","hashOfConfig":"14"},{"size":1198,"mtime":1608519412725,"results":"19","hashOfConfig":"14"},{"size":197,"mtime":1608519251021,"results":"20","hashOfConfig":"14"},{"size":1453,"mtime":1608519766762,"results":"21","hashOfConfig":"14"},{"size":574,"mtime":1608515490254,"results":"22","hashOfConfig":"14"},{"size":1016,"mtime":1608515685828,"results":"23","hashOfConfig":"14"},{"size":1098,"mtime":1608520096241,"results":"24","hashOfConfig":"14"},{"size":696,"mtime":1608515490252,"results":"25","hashOfConfig":"14"},{"filePath":"26","messages":"27","errorCount":0,"warningCount":0,"fixableErrorCount":0,"fixableWarningCount":0},"15srdob",{"filePath":"28","messages":"29","errorCount":0,"warningCount":0,"fixableErrorCount":0,"fixableWarningCount":0,"usedDeprecatedRules":"30"},{"filePath":"31","messages":"32","errorCount":0,"warningCount":2,"fixableErrorCount":0,"fixableWarningCount":0,"source":null},{"filePath":"33","messages":"34","errorCount":0,"warningCount":0,"fixableErrorCount":0,"fixableWarningCount":0},{"filePath":"35","messages":"36","errorCount":0,"warningCount":0,"fixableErrorCount":0,"fixableWarningCount":0},{"filePath":"37","messages":"38","errorCount":0,"warningCount":0,"fixableErrorCount":0,"fixableWarningCount":0,"usedDeprecatedRules":"39"},{"filePath":"40","messages":"41","errorCount":0,"warningCount":0,"fixableErrorCount":0,"fixableWarningCount":0},{"filePath":"42","messages":"43","errorCount":0,"warningCount":2,"fixableErrorCount":0,"fixableWarningCount":0,"source":null},{"filePath":"44","messages":"45","errorCount":0,"warningCount":1,"fixableErrorCount":0,"fixableWarningCount":0,"source":null},{"filePath":"46","messages":"47","errorCount":0,"warningCount":0,"fixableErrorCount":0,"fixableWarningCount":0},{"filePath":"48","messages":"49","errorCount":0,"warningCount":1,"fixableErrorCount":0,"fixableWarningCount":0,"source":null},{"filePath":"50","messages":"51","errorCount":0,"warningCount":2,"fixableErrorCount":0,"fixableWarningCount":0,"source":null},"D:\\project\\spell-card-convert\\spell-card\\src\\App.tsx",[],"D:\\project\\spell-card-convert\\spell-card\\src\\phaser\\scenes\\preloader.js",[],["52","53"],"D:\\project\\spell-card-convert\\spell-card\\src\\phaser\\components\\cardCollection\\cardDisplayZone.js",["54","55"],"D:\\project\\spell-card-convert\\spell-card\\src\\phaser\\components\\navigation.js",[],"D:\\project\\spell-card-convert\\spell-card\\src\\phaser\\scenes\\GameScenes\\cardSwitch.js",[],"D:\\project\\spell-card-convert\\spell-card\\src\\phaser\\components\\gameSetting\\settingMenu.js",[],["56","57"],"D:\\project\\spell-card-convert\\spell-card\\src\\index.tsx",[],"D:\\project\\spell-card-convert\\spell-card\\src\\phaser\\scenes\\GameScenes\\mainGame.js",["58","59"],"D:\\project\\spell-card-convert\\spell-card\\src\\phaser\\scenes\\cardCollectionScenes\\newDeck.js",["60"],"D:\\project\\spell-card-convert\\spell-card\\src\\phaser\\components\\naviButton.js",[],"D:\\project\\spell-card-convert\\spell-card\\src\\components\\SpellCard.jsx",["61"],"D:\\project\\spell-card-convert\\spell-card\\src\\phaser\\scenes\\GameScenes\\deckSelection.js",["62","63"],{"ruleId":"64","replacedBy":"65"},{"ruleId":"66","replacedBy":"67"},{"ruleId":"68","severity":1,"message":"69","line":24,"column":15,"nodeType":"70","messageId":"71","endLine":24,"endColumn":19},{"ruleId":"68","severity":1,"message":"72","line":66,"column":17,"nodeType":"70","messageId":"71","endLine":66,"endColumn":21},{"ruleId":"64","replacedBy":"73"},{"ruleId":"66","replacedBy":"74"},{"ruleId":"68","severity":1,"message":"75","line":2,"column":8,"nodeType":"70","messageId":"71","endLine":2,"endColumn":24},{"ruleId":"68","severity":1,"message":"76","line":37,"column":9,"nodeType":"70","messageId":"71","endLine":37,"endColumn":13},{"ruleId":"68","severity":1,"message":"77","line":18,"column":15,"nodeType":"70","messageId":"71","endLine":18,"endColumn":21},{"ruleId":"68","severity":1,"message":"78","line":11,"column":27,"nodeType":"70","messageId":"71","endLine":11,"endColumn":35},{"ruleId":"68","severity":1,"message":"79","line":18,"column":15,"nodeType":"70","messageId":"71","endLine":18,"endColumn":29},{"ruleId":"68","severity":1,"message":"77","line":20,"column":15,"nodeType":"70","messageId":"71","endLine":20,"endColumn":21},"no-native-reassign",["80"],"no-negated-in-lhs",["81"],"no-unused-vars","'text' is assigned a value but never used.","Identifier","unusedVar","'card' is assigned a value but never used.",["80"],["81"],"'NavigationButton' is defined but never used.","'menu' is assigned a value but never used.","'submit' is assigned a value but never used.","'useState' is defined but never used.","'myColectionBtn' is assigned a value but never used.","no-global-assign","no-unsafe-negation"]
