+++ .gitignore
@@ -1,14 +1,23 @@
-# System and IDE files
-Thumbs.db
+# See https://help.github.com/articles/ignoring-files/ for more about ignoring files.
+
+# dependencies
+/node_modules
+/.pnp
+.pnp.js
+
+# testing
+/coverage
+
+# production
+/build
+
+# misc
 .DS_Store
-.idea
-*.suo
-*.sublime-project
-*.sublime-workspace
+.env.local
+.env.development.local
+.env.test.local
+.env.production.local
 
-# Vendors
-node_modules/
-
-# Build
-dist/
-/npm-debug.log
+npm-debug.log*
+yarn-debug.log*
+yarn-error.log*
