--- index.html
@@ -1,11 +1,0 @@
-<!DOCTYPE html>
-<html lang="en">
-	<head>
-		<meta charset="utf-8" />
-		<title>Phaser 3 React Template</title>
-	</head>
-	<body>
-		<div id="root"></div>
-		<div id="phaser"></div>
-	</body>
-</html>
