+++ src/App.tsx
@@ -1,0 +1,12 @@
+import React from 'react';
+import SpellCard from './components/SpellCard';
+
+const App: React.FC = () => {
+  return (
+    <div className="App">
+        <SpellCard/>
+    </div>
+  );
+}
+
+export default App;
