--- src/index.js
@@ -1,11 +1,0 @@
-import Phaser from "phaser";
-import React from "react";
-import ReactDOM from "react-dom";
-import App from "./components/App.jsx";
-
-//console.log(App);
-
-ReactDOM.render(
-  <App />,
-  document.getElementById("root") || document.createElement("div")
-);
