+++ src/phaser/components/Cards/attackCard.js
@@ -1,11 +1,11 @@
-import Phaser from "phaser";
+import Card from "./Card";
 
-export default class magicCard extends Card {
+export default class attackCard extends Card {
     constructor(name, rank, type, value, imagePath) {
         super(name, rank, type, value, imagePath, isDiscarded)
     }
 
     useEffect(){
-        console.log('magic damage is' + value);
+        console.log('attack damage is' + value);
     }
 }
