+++ src/phaser/components/Cards/defenseCard.js
@@ -1,5 +1,4 @@
-import Phaser from "phaser";
-
+import Card from './Card'
 export default class defenseCard extends Card {
     constructor(name, rank, type, value, imagePath) {
         super(name, rank, type, value, imagePath, isDiscarded)
