+++ src/phaser/components/Cards/magicCard.js
@@ -1,0 +1,11 @@
+import Card from './Card'
+
+export default class magicCard extends Card {
+    constructor(name, rank, type, value, imagePath) {
+        super(name, rank, type, value, imagePath, isDiscarded)
+    }
+
+    useEffect(){
+        console.log('magic damage is' + value);
+    }
+}
