+++ src/phaser/components/cardGenerator.js
