+++ src/phaser/components/zone.js
