+++ src/phaser/scenes/GameScenes/gamePreloader.js
@@ -1,3 +1,5 @@
+import Phaser from 'phaser'
+
 export default class gamePreloader extends Phaser.Scene{
     constructor(){
         super("gamePreload")
