--- src/phaser/spellCard/components/Cards/attackCard.js
@@ -1,12 +1,0 @@
-import Phaser from "phaser";
-import Card from "./Card";
-
-export default class attackCard extends Card {
-    constructor(name, rank, type, value, imagePath) {
-        super(name, rank, type, value, imagePath, isDiscarded)
-    }
-
-    useEffect(){
-        console.log('attack damage is' + value);
-    }
-}
